Make solvers crate usable as library
