+++ crates/solvers/src/lib.rs
@@ -1,0 +1,12 @@
+// TODO remove this once the crate stabilizes a bit.
+#![allow(dead_code)]
+
+mod api;
+pub mod boundary;
+mod cli;
+mod config;
+mod domain;
+pub mod run;
+#[cfg(test)]
+mod tests;
+mod util;
